# src/client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LemonDo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfTaskDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/uncomplete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/archive": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/move": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/tags": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/tags/{tag}": {
      "delete": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/{id}/view-note": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewTaskNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tasks/bulk/complete": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddTagRequest": {
        "required": [
          "tag"
        ],
        "type": "object",
        "properties": {
          "tag": {
            "type": "string"
          }
        },
        "description": "Adds a tag to a task with automatic normalization."
      },
...
      "BulkCompleteRequest": {
        "required": [
          "taskIds"
        ],
        "type": "object",
        "properties": {
          "taskIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "description": "Completes multiple tasks in a single operation."
      },
...
      "CreateTaskRequest": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "enum": [
              "None",
              "Low",
              "Medium",
              "High",
              "Critical"
            ],
            "type": "string",
            "default": "None"
          },
          "dueDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "sensitiveNote": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Creates a new task with optional metadata."
      },
...
      "MoveTaskRequest": {
        "required": [
          "columnId",
          "previousTaskId",
          "nextTaskId"
        ],
        "type": "object",
        "properties": {
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "previousTaskId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "nextTaskId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        },
        "description": "Moves a task to a new position using neighbor-based ranking."
      },
...
      "PagedResultOfTaskDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Total number of pages available based on int PagedResult&lt;T&gt;.TotalCount and int PagedResult&lt;T&gt;.PageSize.",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "True if there are more pages after the current int PagedResult&lt;T&gt;.Page."
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "True if there are pages before the current int PagedResult&lt;T&gt;.Page. False when on the first page."
          }
        },
        "description": "Represents a paginated result set."
      },
...
      "TaskDto": {
        "required": [
          "id",
          "title",
          "priority",
          "status",
          "tags",
          "isArchived",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the task.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Title or summary of the task."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "Detailed description or notes for the task."
          },
          "priority": {
            "enum": [
              "None",
              "Low",
              "Medium",
              "High",
              "Critical"
            ],
            "type": "string",
            "description": "Priority level (None, Low, Medium, High)."
          },
          "status": {
            "enum": [
              "Todo",
              "InProgress",
              "Done"
            ],
            "type": "string",
            "description": "Current lifecycle status (Todo, InProgress, Done)."
          },
          "dueDate": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional deadline for task completion.",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Collection of tags for categorizing or labeling the task."
          },
          "isArchived": {
            "type": "boolean",
            "description": "Indicates if the task is archived (hidden from active views but not deleted)."
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Indicates if the task has been soft-deleted."
          },
          "sensitiveNote": {
            "type": [
              "null",
              "string"
            ],
            "description": "Whether this task has an encrypted sensitive note. Shows \"[PROTECTED]\" or null."
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "Timestamp when the task was marked as Done, or null if not completed.",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the task was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when the task was last modified.",
            "format": "date-time"
          }
        },
        "description": "Read model for a task, flattening value objects to primitive types for API responses."
      },
...
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "priority": {
            "enum": [
              "None",
              "Low",
              "Medium",
              "High",
              "Critical"
            ],
            "type": [
              "null",
              "string"
            ]
          },
          "dueDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "clearDueDate": {
            "type": "boolean",
            "default": false
          },
          "sensitiveNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "clearSensitiveNote": {
            "type": "boolean",
            "default": false
          }
        },
        "description": "Updates a task with partial field modifications."
      },
...
      "ViewTaskNoteRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "description": "Views a task's sensitive note via break-the-glass authentication."
      }
    }
  },
...
}